Readme & usage text changes, more hyphens less underscores
Previously the readme was exactly the usage text, but some
differences make sense. I made the readme markdown, and
removed some of the high-level description and summary from
the usage but left it in the readme

I made the readme markdown. and removed some of the content from
the usage text but left it in the readme

Also, changed underscores in args and commands to be hyphens
which is more standard